pyproject.toml
[project]
name = "textverified"
version = "0.1.0a0"
description = "Python wrapper for the TextVerified API"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Textverified" },
]

keywords = ["Textverified", "Verification", "Web Scraping", "API", "webhook", "python"]

dependencies = ["requests>=2.31.0", "urllib3>=1.26.20", "python-dateutil>=2.9.0"]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "networkx>=3.1",
    "build>=1.1.1"
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]

[project.urls]
Homepage = "https://textverified.com"
Download = "https://pypi.org"

[tool.black]
line-length = 120
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.venv
  | build
  | dist
  | textverified/generated
)/
'''

[tool.pytest.ini_options]
addopts = """
--verbose \
--cov=textverified \
--cov-report=xml"""
python_files = ["tests/*.py"]

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"